+++ libs/cardano-ledger-core/CHANGELOG.md
@@ -8,6 +8,7 @@
 * Add `upgradeScript` function to `EraScript`
 * Add `upgradeTxAuxData` function to `EraTxAuxData`
 * Add `upgradeTxCert` function and `TxCertUpgradeError` family to `EraTxCert`
+* Add `Default` instance for `Anchor`
 * Add `Anchor` and `AnchorDataHash`
 * Add `DRepState`
 * Change `vsDReps` to a map